updated models types
