--- server/db/models/customer.js
@@ -1,37 +1,0 @@
-const Sequelize = require('sequelize')
-const db = require('../db')
-
-const Customer = db.define('customer', {
-  email: {
-    type: Sequelize.STRING,
-    unique: true,
-    allowNull: false,
-    validate: {
-      notEmpty: true
-    }
-  },
-  password: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    validate: {
-      len: [10, 100],
-      notEmpty: true
-    }
-  },
-  firstName: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    validate: {
-      notEmpty: true
-    }
-  },
-  lastName: {
-    type: Sequelize.STRING,
-    allowNull: false,
-    validate: {
-      notEmpty: true
-    }
-  }
-})
-
-module.exports = Customer
