+++ server/db/models/products.js
@@ -17,7 +17,7 @@ const Product = db.define('products', {
     }
   },
   price: {
-    type: Sequelize.FLOAT,
+    type: Sequelize.INTEGER,
     allowNull: false,
     validate: {
       notEmpty: true,
